api/requirements.txt
Django==3.2.16
djangorestframework
debugpy
isort
poetry
mypy
black
pyproject-flake8
psycopg2
djangorestframework-jwt
django-safedelete